CP-54351 move lvmanager from /opt
